feat: finished the review endpoint currently and update something ..

+++ Bookistry.API/Entities/Review.cs
@@ -3,7 +3,7 @@
 public sealed class Review : Auditable
 {
     public Guid Id { get; set; } = Guid.NewGuid();
-    public double Rating { get; set; } // rating for every user
+    public int Rating { get; set; } 
     public string Comment { get; set; } = string.Empty;
 
     public Guid BookId { get; set; }
